automatic_ticket_classification_tool/requirements.txt
streamlit
pypdf
langchain
langchain-openai
langchain-pinecone
langchain-community
python-dotenv
scikit-learn
pandas
joblib